requirements.txt
requests
pywebview
ruamel.yaml